dnfremove

+++ .config/fish/functions/dnfremove.fish
@@ -3,5 +3,10 @@ function dnfremove --wraps='dnf install'
     for arg in $argv
         filterfile ~/.github/dnf_installed $arg
     end
-    sudo dnf remove $argv
+    if command -v rpm-ostree >/dev/null
+        sudo rpm-ostree remove $argv
+        echo sudo rpm-ostree override remove
+    else
+        sudo dnf remove $argv
+    end
 end
